refactor(native): separate browser and account bridge endpoints
Extract browser callback adaptation, gateway endpoints, account endpoints, provider connection operations, and callback-scoped text marshalling without changing the C ABI.
