+++ packages/agent-native-bridge/test/Spec.hs
@@ -12,6 +12,7 @@ import qualified Agent.CLI.MacOS.TaskSchedulerSpec as TaskSchedulerSpec
 import qualified Agent.CLI.McpAdminSpec as McpAdminSpec
 import qualified Agent.CLI.ResourceAdminSpec as ResourceAdminSpec
 #ifdef darwin_HOST_OS
+import qualified Agent.CLI.MacOS.AccountConnectionSpec as AccountConnectionSpec
 import qualified Agent.CLI.MacOS.BridgeFFISpec as BridgeFFISpec
 import qualified Agent.CLI.MacOS.BridgeHeaderSpec as BridgeHeaderSpec
 import qualified Agent.CLI.MacOS.BridgeSpec as BridgeSpec
@@ -29,6 +30,7 @@ main = hspec do
     TaskSchedulerSpec.spec
     RepositoryWorkersSpec.spec
 #ifdef darwin_HOST_OS
+    AccountConnectionSpec.spec
     BrowserBridgeFFISpec.spec
     BridgeFFISpec.spec
     BridgeHeaderSpec.spec
